+++ debug/debug.go
@@ -3,34 +3,17 @@ package debug
 import (
 	"bufio"
 	"bytes"
-	"errors"
 	"fmt"
 	"github.com/dragonsinth/gaddis"
 	"github.com/dragonsinth/gaddis/asm"
 	"github.com/dragonsinth/gaddis/ast"
 	"github.com/dragonsinth/gaddis/gogen/builtins"
-	"log"
 	"math/rand"
 	"os"
 	"sync"
 	"sync/atomic"
 	"time"
 )
-
-type EventHost interface {
-	Paused(reason string)
-	Exception(err error)
-	Panicked(error, []ast.Position, []string)
-	Exited(code int)
-	Terminated()
-}
-
-type Opts struct {
-	IsTest     bool
-	Stdin      []byte
-	Stdout     func(string)
-	WorkingDir string // TODO
-}
 
 type Session struct {
 	Host      EventHost
@@ -66,28 +49,6 @@ type Session struct {
 	stepLine    int
 	stepFrame   int
 }
-
-type RunState int
-
-const (
-	// keep running
-	RUN RunState = iota
-	// halt with a pause event
-	PAUSE
-	// halt with no event
-	HALT
-	// halt with a terminate event
-	TERMINATE
-)
-
-type StepType int
-
-const (
-	STEP_NONE StepType = iota
-	STEP_NEXT
-	STEP_IN
-	STEP_OUT
-)
 
 func New(
 	filename string,
@@ -148,13 +109,6 @@ func New(
 	return ds, nil
 }
 
-func (ds *Session) MapSourceLine(i int) int {
-	if i < 0 || i >= ds.NLines {
-		return -1
-	}
-	return ds.SourceToInst[i]
-}
-
 func (ds *Session) Reset(opts Opts) {
 	var seed int64
 	if !opts.IsTest {
@@ -164,8 +118,9 @@ func (ds *Session) Reset(opts Opts) {
 	ec := &asm.ExecutionContext{
 		Rng: rand.New(rand.NewSource(seed)),
 		IoContext: builtins.IoContext{
-			Stdin:  bufio.NewScanner(bytes.NewReader(opts.Stdin)),
-			Stdout: &bufferedSyncWriter{out: opts.Stdout},
+			Stdin:   bufio.NewScanner(bytes.NewReader(opts.Stdin)),
+			Stdout:  &bufferedSyncWriter{out: opts.Stdout},
+			WorkDir: opts.WorkDir,
 		},
 	}
 
@@ -184,241 +139,4 @@ func (ds *Session) Reset(opts Opts) {
 		ds.instBreaks = make([]byte, ds.NInst)
 	}
 	ds.stepType = STEP_NONE
-}
-
-func (ds *Session) Play() {
-	if !ds.running.CompareAndSwap(false, true) {
-		return
-	}
-
-	ds.runMu.Lock()
-	defer ds.runMu.Unlock()
-	if ds.runState > PAUSE {
-		log.Println("ERROR: invalid runstate:", ds.runState)
-		return
-	}
-	if ds.isDone {
-		log.Println("ERROR: already done")
-		return
-	}
-
-	ds.runState = RUN
-	log.Println("running")
-
-	go func() {
-		defer ds.running.Store(false)
-		p := ds.Exec
-		ds.runMu.Lock()
-		defer ds.runMu.Unlock()
-
-		defer func() {
-			// if we exit for any reason, reset all step state
-			ds.stepType = STEP_NONE
-
-			if r := recover(); r != nil {
-				err := errors.New(fmt.Sprint(r))
-				log.Println("panicking:", err)
-				if ds.noDebug {
-					// execute the panic; send trace to stderr
-					var lines []ast.Position
-					var trace []string
-					ds.Exec.GetStackFrames(func(fr *asm.Frame, _ int, inst asm.Inst) {
-						lines = append(lines, inst.GetSourceInfo().Start)
-						trace = append(trace, asm.FormatFrameScope(fr))
-					})
-					ds.Host.Panicked(err, lines, trace)
-					ds.isDone = true
-				} else {
-					// stop on exception
-					ds.runState = PAUSE
-					ds.exception = err
-					ds.Host.Exception(err)
-				}
-			}
-		}()
-
-		runStateEvent := func() {
-			switch ds.runState {
-			case HALT:
-				log.Println("halting")
-			case PAUSE:
-				ds.Host.Paused("pause")
-				log.Println("pausing")
-			case TERMINATE:
-				ds.Host.Terminated()
-				log.Println("terminating")
-			default:
-				panic(ds.runState)
-			}
-		}
-
-		if ds.runState != RUN {
-			runStateEvent()
-			return
-		}
-
-		if ds.stopOnEntry {
-			ds.stopOnEntry = false // just once
-			ds.Host.Paused("entry")
-			return
-		}
-
-		instructionCount := 0
-		for p.Frame != nil {
-			// someone is asking for the lock; eventually we'll yield it
-			if ds.yield.Load() {
-				func() {
-					ds.runMu.Unlock()
-					defer ds.runMu.Lock()
-				}()
-
-				if ds.runState != RUN {
-					runStateEvent()
-					return
-				}
-			}
-
-			inst := p.Code[p.PC]
-			if ds.stepType != STEP_NONE {
-				stackDiff := len(p.Stack) - ds.stepFrame
-				line := inst.GetSourceInfo().Start.Line
-				switch ds.stepType {
-				case STEP_NEXT:
-					if stackDiff < 0 || (stackDiff == 0 && ds.stepLine != line) {
-						ds.Host.Paused("step")
-						ds.runState = PAUSE
-						return
-					}
-				case STEP_IN:
-					// break on any change
-					if stackDiff != 0 || ds.stepLine != line {
-						ds.Host.Paused("step")
-						ds.runState = PAUSE
-						return
-					}
-				case STEP_OUT:
-					if stackDiff < 0 {
-						ds.Host.Paused("step")
-						ds.runState = PAUSE
-						return
-					}
-				default:
-					panic(ds.stepType)
-				}
-			}
-
-			inst.Exec(p)
-			p.PC++
-
-			// must be after exec to avoid reentry
-			if ds.instBreaks[p.PC] != 0 {
-				ds.Host.Paused("breakpoint")
-				ds.runState = PAUSE
-				return
-			}
-
-			instructionCount++
-			if instructionCount > asm.MAX_INSTRUCTIONS {
-				panic("infinite loop detected")
-			}
-		}
-		ds.isDone = true
-		ds.Host.Exited(0)
-	}()
-}
-
-func (ds *Session) Terminate() {
-	ds.yield.Store(true)
-	ds.runMu.Lock()
-	defer ds.runMu.Unlock()
-	ds.yield.Store(false)
-	ds.runState = TERMINATE
-}
-
-func (ds *Session) Pause() {
-	ds.yield.Store(true)
-	ds.runMu.Lock()
-	defer ds.runMu.Unlock()
-	ds.yield.Store(false)
-	ds.runState = PAUSE
-}
-
-// Halt without sending any events. Wait for halt.
-func (ds *Session) Halt() {
-	ds.withOuterLock(func() {
-		ds.runState = HALT
-	})
-	for ds.running.Load() {
-		ds.withOuterLock(func() {})
-	}
-}
-
-func (ds *Session) SetNoDebug() {
-	ds.withOuterLock(func() {
-		ds.noDebug = true
-	})
-}
-
-func (ds *Session) SetBreakpoints(bps []int) {
-	ds.withOuterLock(func() {
-		clear(ds.instBreaks)
-		clear(ds.lineBreaks)
-		for _, bp := range bps {
-			if bp < 0 || bp >= len(ds.lineBreaks) {
-				continue
-			}
-			pc := ds.SourceToInst[bp]
-			if pc < 0 {
-				continue
-			}
-			ds.lineBreaks[bp] = 1
-			ds.instBreaks[pc] = 1
-		}
-	})
-}
-
-func (ds *Session) StopOnEntry() {
-	ds.withOuterLock(func() {
-		ds.stopOnEntry = true
-	})
-}
-
-func (ds *Session) GetStackFrames(f func(*asm.Frame, int, asm.Inst)) {
-	ds.withOuterLock(func() {
-		ds.Exec.GetStackFrames(f)
-	})
-}
-
-func (ds *Session) Step(stepType StepType) {
-	ds.withOuterLock(func() {
-		p := ds.Exec
-		si := p.Code[p.PC].GetSourceInfo()
-		ds.stepType = stepType
-		ds.stepLine = si.Start.Line
-		ds.stepFrame = len(p.Stack)
-	})
-}
-
-func (ds *Session) RestartFrame(id int) {
-	ds.withOuterLock(func() {
-		p := ds.Exec
-		if id < 1 || id > len(p.Stack) {
-			return // client error
-		}
-		p.Stack = p.Stack[:id]
-		p.Frame = &p.Stack[id-1]
-		p.PC = p.Frame.Start
-		clear(p.Frame.Locals)
-		copy(p.Frame.Locals, p.Frame.Args)
-		p.Frame.Eval = p.Frame.Eval[:0]
-	})
-}
-
-func (ds *Session) withOuterLock(f func()) {
-	// force a yield, acquire the lock
-	ds.yield.Store(true)
-	ds.runMu.Lock()
-	defer ds.runMu.Unlock()
-	ds.yield.Store(false)
-	f()
 }
