+++ examples/examples.go
@@ -120,8 +120,9 @@ func RunTestInterp(t *testing.T, filename string) error {
 	p := cp.NewExecution(&asm.ExecutionContext{
 		Rng: rand.New(rand.NewSource(0)),
 		IoContext: builtins.IoContext{
-			Stdin:  bufio.NewScanner(&input),
-			Stdout: gaddis.NoopSyncWriter(&output),
+			Stdin:   bufio.NewScanner(&input),
+			Stdout:  gaddis.NoopSyncWriter(&output),
+			WorkDir: ".",
 		},
 	})
